add recon scripts and config files
