--- gadgetron-python-demo/recon_radial.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration>
-    <version>2</version>
-
-    <readers>
-        <reader>
-            <dll>gadgetron_core_readers</dll>
-            <classname>AcquisitionReader</classname>
-        </reader>
-        <reader>
-            <dll>gadgetron_core_readers</dll>
-            <classname>WaveformReader</classname>
-        </reader>
-    </readers>
-
-    <writers>
-        <writer>
-            <dll>gadgetron_core_writers</dll>
-            <classname>ImageWriter</classname>
-        </writer>
-    </writers>
-
-    <stream>
-        <external>
-            <execute name="dummy_conn_recon_acquisitions" target="recon" type="python"/>
-...
-            <configuration/>
-        </external>
-    </stream>
-</configuration>
